Update Senior_Captsone.py
I added more code for pulling tweets from both movies and then applied sentiment analysis to the tweets. This is just a base case. We have a lot of cleaning to do of the data, especially now that 'parasite' has become a verb related to Coronavirus. Still about the movie, but people are being 'parasited' into their homes.